AI-Agents/Agent-06-handoff/pyproject.toml
[project]
name = "agent-06-handoff"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
